pyproject.toml
[tool.poetry]
authors = ["Jérémie Lumbroso <lumbroso@seas.upenn.edu>"]
description = "IntegriMark is a system that protects digital PDF documents by watermarking them with user-specific details, ensuring traceability and discouraging unauthorized sharing."
include = ["integrimark.pack.html"]
license = "LPGLv3"
name = "integrimark"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
click = "^8.1.7"
loguru = "^0.7.2"
pycryptodome = "^3.19.0"
python = "^3.11"
xkcdpass = "^1.19.4"
gspread = "^5.12.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
